SessionTracker with Timer implemented. With this tracker you can start and finish session.
