+++ Comptech.Backend/src/Comptech.Backend.Data/DomainEntities/SessionStatus.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Comptech.Backend.Data.DomainEntities
+{
+    public enum SessionStatus
+    {
+        ACTIVE,
+        FINISHED
+    }
+}
