--- Comptech.Backend/src/Comptech.Backend.Service/Core/Sessions/ValidateSessionAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Comptech.Backend.Service.Core.Sessions
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public class ValidateSessionAttribute : Attribute
-    {
-    }
-}
