--- Comptech.Backend/src/Comptech.Backend.Service/Services/Security/CryptographyService.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Comptech.Backend.Service.Services.Security
-{
-    public interface ICryptographyService
-    {
-        byte[] Decrypt(byte[] encryptedImage);
-    }
-
-    public class CryptographyService : ICryptographyService
-    {
-        public byte[] Decrypt(byte[] encryptedImage)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
